+++ src/parser.js
@@ -39,12 +39,12 @@ function parseTextWithSpawn(text, anonymous) {
   return executionResult.stdout.toString();
 }
 
-function parseTextWithHttp(text, serverPort, anonymous) {
+function parseTextWithHttp(text, serverHost, serverPort, anonymous) {
   const httpClientLocation = path.join(__dirname, "http-client.js");
   const args = [
     httpClientLocation,
     "-a",
-    "localhost",
+    serverHost,
     "-f",
     "json",
     "-p",
@@ -481,6 +481,7 @@ function parse(sourceCode, _, options) {
   if (options.apexStandaloneParser === "built-in") {
     serializedAst = parseTextWithHttp(
       sourceCode,
+      options.apexStandaloneHost,
       options.apexStandalonePort,
       options.parser === "apex-anonymous",
     );