+++ tests_config/run_spec.js
@@ -13,6 +13,7 @@ function prettyPrint(src, filename, options) {
     filepath: filename,
     apexStandaloneParser: "built-in",
     apexStandalonePort: 2117,
+    apexStandaloneHost: "localhost",
     ...options,
   });
 }
@@ -24,6 +25,7 @@ function parse(string, opts) {
     {
       apexStandaloneParser: "built-in",
       apexStandalonePort: 2117,
+      apexStandaloneHost: "localhost",
       ...opts,
     },
     /* massage */ true,